refactor: removes connection from workspaces
